refactor: auth and store data now split into separte files to reduce merge conflicts from folks putting imports in different orders when working on other branches.
