+++ src/firebase/app.js
@@ -1,0 +1,6 @@
+import { initializeApp } from 'firebase/app';
+import firebaseConfig from './firebaseConfig';
+
+const app = initializeApp(firebaseConfig);
+
+export default app;
