+++ src/firebase/database.js
@@ -1,19 +1,6 @@
-import { initializeApp } from 'firebase/app';
-import { getAuth, sendPasswordResetEmail, onAuthStateChanged,signOut } from 'firebase/auth';
+import { getFirestore, serverTimestamp } from 'firebase/firestore';
+import app from './firebaseApp';
 
-import{ getFirestore,serverTimestamp } from 'firebase/firestore';
+const firebaseStore = getFirestore(app);
 
-const firebaseConfig = {
-  apiKey: import.meta.env.VITE_API_KEY,
-  authDomain: import.meta.env.VITE_AUTH_DOMAIN,
-  projectId: import.meta.env.VITE_PROJECT_ID,
-  storageBucket: import.meta.env.VITE_STORAGE_BUCKET,
-  messagingSenderId: import.meta.env.VITE_MESSAGE_SENDER_ID,
-  appId: import.meta.env.VITE_APP_ID
-};
-
-const app = initializeApp(firebaseConfig);
-const firebaseAuth = getAuth(app);
-const firebaseStore = getFirestore();
-
-export{ app,firebaseAuth,firebaseStore,serverTimestamp,sendPasswordResetEmail , onAuthStateChanged,signOut };
+export { firebaseStore, serverTimestamp };
