+++ src/firebase/firebaseAuth.js
@@ -1,0 +1,10 @@
+import { getAuth, sendPasswordResetEmail, onAuthStateChanged, signOut } from 'firebase/auth';
+import app from './firebaseApp';
+
+const firebaseAuth = getAuth(app);
+
+export {
+  firebaseAuth,
+  sendPasswordResetEmail,
+  onAuthStateChanged,
+  signOut };
